refactor: extract editor pickers
